Extract BaseCodeModeRuntime from IsolatedVmCodeModeRuntime
- Create src/runtime/base-runtime.ts with shared logic (tool indexing,
  host-side tool calling, search, schema, logging)
- Refactor IsolatedVmCodeModeRuntime to extend BaseCodeModeRuntime
- Update createCodeModeRuntime to accept runtime option
- Export BaseCodeModeRuntime from package
